bugfixed LT in simulator added new LT2 test in simple_tests.cpp added sad16 test with both unrolled and jmp for ConnexS. Works ok in simulator. Untested on HW. Added two targets for arm in both opincaa-lib and connexppAsm
git-svn-id: https://apollo.arh.pub.ro:8443/svn/ConnexSIFT/trunk/software/connex@325 0e9f8c0c-05a5-1044-940b-b810a0af9d24

+++ opincaa/old/connexppIAsmBranch/src/c_simu/c_simulator.cpp
@@ -403,7 +403,7 @@ int c_simulator::DeAsmBatch(UINT16 dwBatchNumber)
                                                    continue;}
                 case _LT:
                     {
-                        FOR_ALL_ACTIVE_MACHINES( if(CSimuRegs[MACHINE][GET_LEFT(CI)] < CSimuRegs[MACHINE][GET_RIGHT(CI)])
+                        FOR_ALL_ACTIVE_MACHINES( if((INT16)CSimuRegs[MACHINE][GET_LEFT(CI)] < (INT16)CSimuRegs[MACHINE][GET_RIGHT(CI)])
                                                     { CSimuRegs[MACHINE][GET_DEST(CI)] = 1; CSimuLtFlags[MACHINE] = 1; }
                                                 else { CSimuRegs[MACHINE][GET_DEST(CI)] = 0; CSimuLtFlags[MACHINE] = 0; };  );
                         continue;}
